requirements.txt
fastapi
uvicorn[standard]
cohere
python-multipart
pydantic
numpy
python-dotenv
requests
scikit-learn